Pesquisa por Autor (em andamento)

+++ src/br/livraria/dao/PesquisaDao.java
@@ -9,7 +9,7 @@ import br.livraria.dominio.Livro;
 
 public interface PesquisaDao {
 	
-	public List<Livro> pesquisarPorAutor(Integer codigoAutor) throws SQLException;
+	public List<Livro> pesquisarPorAutor(Long codigoAutor) throws SQLException;
 	public List<Livro> pesquisarPorTitulo(String titulo) throws SQLException;
 	public List<Livro> pesquisarPorCategoria(Long codigoCategoria) throws SQLException;
